feat: 의존성 역전 수정
